Add toggle for experimental OpenGL ES 3.0 support
This adds a toggle for experimental OpenGL ES 3.0 support to the
developer settings. On Android 7 for FP2 we had to fall back to
OpenGL ES 2.0 because of failing CTS tests. This however caused
compatibility regressions with 3D apps such as Pokemon Go. Since these
apps do work with OpenGL ES 3.0 enabled, we introduce a toggle allowing
users to enable it by hand. For now we leave it as an experimental
feature in developer settings, so that we can collect data on its
behavior.

Issue: FP2N-301
Change-Id: I140295801d7adaa5c69fd170642948a44b043aa8
Depends-On: I22d18d7f247207d0f0dd04441e77381b4fce113e